Document imported functions from mlflow.pyfunc.model, rather than wrapping a _save_model method
