+++ docs/source/python_api/mlflow.pyfunc.rst
@@ -6,16 +6,19 @@ mlflow.pyfunc
     :undoc-members:
     :show-inheritance:
 
+    .. autofunction:: mlflow.pyfunc.save_model
+    .. autofunction:: mlflow.pyfunc.log_model
+
 .. Include ``PythonModelContext`` as a renamed class to avoid documenting constructor parameters.
    This class is meant to be constructed implicitly, and users should only be aware of its
    documented member properties.
-.. autoclass:: mlflow.pyfunc.model.PythonModelContext()
+.. autoclass:: mlflow.pyfunc.PythonModelContext()
     :members:
     :undoc-members:
 
-.. Include ``PythonModel`` as a class rather than using "automodule" to include documentation for
-   the ``__init__`` method without documenting ``__init__`` for other classes in the same module.
-.. autoclass:: mlflow.pyfunc.model.PythonModel
+.. Include ``PythonModel`` as a class to include documentation for the ``__init__`` method
+   without documenting ``__init__`` for other classes in the same module.
+.. autoclass:: mlflow.pyfunc.PythonModel
     :members:
     :undoc-members:
